# nest-backend/openApiSpec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/tenants/by-origin": {
            "get": {
                "operationId": "TenantController_findTenantByOrigin",
                "parameters": [
                    {
                        "name": "origin",
                        "required": true,
                        "in": "header",
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
        "/tenants": {
            "get": {
                "operationId": "TenantController_findOneTenant",
                "parameters": [
                    {
                        "name": "name",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "host",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "themeId",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            },
            "post": {
                "operationId": "TenantController_createTenant",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTenantAppDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
        "/tenants/{id}/workspaces": {
            "get": {
                "operationId": "TenantController_getWorkspaces",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
        "/tenants/cms": {
            "post": {
                "operationId": "TenantController_createTenantCMS",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTenantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
        "/tenants/cms/{id}": {
            "patch": {
                "operationId": "TenantController_updateTenantCMS",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTenantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
        "/tenants/{id}": {
            "patch": {
                "operationId": "TenantController_updateTenant",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTenantFromWebAppDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Tenants"]
            }
        },
...
        "/workspaces/{id}/people/{userId}": {
            "get": {
                "operationId": "WorkspacePeopleController_findOneUser",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Workspaces - People"]
            }
        },
        "/workspaces/{id}/people": {
            "get": {
                "operationId": "WorkspacePeopleController_findPeople",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "tags": ["Workspaces - People"]
            }
        },
        "/workspaces/{id}/recommended/{userId}": {
            "get": {
                "operationId": "WorkspacePeopleController_recommendPeer",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SanitizedExpertDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Workspaces - People"]
            }
        },
        "/workspaces/{id}/message/{userId}": {
            "post": {
                "operationId": "WorkspacePeopleController_sendMessage",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendMessageDto"
                            }
                        }
                    }
                },
                "responses": { "201": { "description": "" } },
                "tags": ["Workspaces - People"]
            }
        },
        "/workspaces": {
            "post": {
                "operationId": "create_one_workspace",
                "summary": "",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateWorkspaceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Workspaces"]
            },
            "get": {
                "operationId": "find_many_workspaces",
                "summary": "",
                "parameters": [
                    {
                        "name": "tenantId",
                        "required": true,
                        "in": "query",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "search",
                        "required": true,
                        "in": "query",
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "tags": ["Workspaces"]
            }
        },
...
    },
    "info": {
        "title": "Posters example",
        "description": "The posters API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "Upload": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "fieldname": { "type": "string" },
                    "originalname": { "type": "string" },
                    "encoding": { "type": "string" },
                    "contentEncoding": { "type": "string" },
                    "mimetype": { "type": "string" },
                    "size": { "type": "number" },
                    "bucket": { "type": "string" },
                    "key": { "type": "string" },
                    "acl": { "type": "string" },
                    "contentType": { "type": "string" },
                    "contentDisposition": { "type": "string" },
                    "storageClass": { "type": "string" },
                    "serverSideEncryption": { "type": "string" },
                    "metadata": { "type": "object" },
                    "location": { "type": "string" },
                    "etag": { "type": "string" },
                    "versionId": { "type": "string" }
                },
                "required": [
                    "id",
                    "fieldname",
                    "originalname",
                    "encoding",
                    "contentEncoding",
                    "mimetype",
                    "size",
                    "bucket",
                    "key",
                    "acl",
                    "contentType",
                    "contentDisposition",
                    "storageClass",
                    "serverSideEncryption",
                    "metadata",
                    "location",
                    "etag",
                    "versionId"
                ]
            },
...
            "TenantSuperadmin": {
                "type": "object",
                "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" }
                },
                "required": ["firstName", "lastName", "email"]
            },
            "TenantTheme": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "tenantId": { "type": "string" },
                    "name": { "type": "string" },
                    "host": { "type": "string" },
                    "domain": { "type": "string" },
                    "enabled": { "type": "boolean" },
                    "superadmin": {
                        "$ref": "#/components/schemas/TenantSuperadmin"
                    },
                    "features": { "type": "object" },
                    "images": { "type": "object" },
                    "css": { "type": "object" }
                },
                "required": [
                    "name",
                    "host",
                    "domain",
                    "enabled",
                    "superadmin",
                    "features",
                    "images",
                    "css"
                ]
            },
            "Tenant": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "name": { "type": "string" },
                    "host": { "type": "string" },
                    "roles": { "type": "array", "items": { "type": "object" } },
                    "themeId": { "type": "number" },
                    "logo": { "$ref": "#/components/schemas/Upload" },
                    "logoId": { "type": "number", "nullable": true },
                    "ownerId": { "type": "number" },
                    "termOfUse": { "type": "boolean" },
                    "enabled": { "type": "boolean" },
                    "features": {
                        "$ref": "#/components/schemas/TenantFeatures"
                    },
                    "theme": { "$ref": "#/components/schemas/TenantTheme" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" },
                    "domain": { "type": "string", "nullable": true }
                },
                "required": [
                    "id",
                    "name",
                    "host",
                    "themeId",
                    "logoId",
                    "ownerId",
                    "termOfUse",
                    "enabled",
                    "createdAt",
                    "updatedAt",
                    "domain"
                ]
            },
            "TenantFeatures": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "AccessControl": { "type": "boolean" },
                    "TenantManagement": { "type": "boolean" },
                    "Workspace": { "type": "boolean" },
                    "Project": { "type": "boolean" },
                    "Gallery": { "type": "boolean" },
                    "People": { "type": "boolean" },
                    "Comment": { "type": "boolean" },
                    "tenant": { "$ref": "#/components/schemas/Tenant" }
                },
                "required": [
                    "AccessControl",
                    "TenantManagement",
                    "Workspace",
                    "Project",
                    "Gallery",
                    "People",
                    "Comment"
                ]
            },
            "CreateTenantDto": {
                "type": "object",
                "properties": {
                    "superadmin": {
                        "$ref": "#/components/schemas/TenantSuperadmin"
                    },
                    "name": { "type": "string" },
                    "host": { "type": "string" },
                    "features": {
                        "$ref": "#/components/schemas/TenantFeatures"
                    },
                    "themeId": { "type": "number" },
                    "enabled": { "type": "boolean" },
                    "logo": { "$ref": "#/components/schemas/Upload" },
                    "domain": { "type": "string", "nullable": true }
                },
                "required": [
                    "superadmin",
                    "name",
                    "host",
                    "themeId",
                    "enabled",
                    "domain"
                ]
            },
            "UpdateTenantDto": {
                "type": "object",
                "properties": {
                    "superadmin": {
                        "$ref": "#/components/schemas/TenantSuperadmin"
                    },
                    "name": { "type": "string" },
                    "host": { "type": "string" },
                    "features": {
                        "$ref": "#/components/schemas/TenantFeatures"
                    },
                    "enabled": { "type": "boolean" },
                    "domain": { "type": "string", "nullable": true }
                }
            },
            "CreateTenantAppDto": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "host": { "type": "string" },
                    "logo": { "$ref": "#/components/schemas/Upload" },
                    "domain": { "type": "string", "nullable": true }
                },
                "required": ["name", "host", "domain"]
            },
            "UpdateTenantFromWebAppDto": {
                "type": "object",
                "properties": { "name": { "type": "string" } },
                "required": ["name"]
            },
...
            "SanitizedExpertDto": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "objectId": { "type": "string" },
                    "organizationName": { "type": "string", "nullable": true },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "isDummyUser": { "type": "boolean" },
                    "disableNotifications": { "type": "boolean" },
                    "linkedin": { "type": "string" },
                    "researchGate": { "type": "string" },
                    "twitter": { "type": "string" },
                    "currentPosition": { "type": "string", "nullable": true },
                    "biography": { "type": "string", "nullable": true },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" },
                    "avatarId": { "type": "number", "nullable": true },
                    "avatar": { "$ref": "#/components/schemas/Upload" },
                    "authoredPosters": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "roles": { "type": "array", "items": { "type": "object" } },
                    "comments": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "userDirectory": { "type": "object" },
                    "skills": { "type": "array", "items": { "type": "object" } }
                },
                "required": [
                    "id",
                    "objectId",
                    "organizationName",
                    "firstName",
                    "lastName",
                    "email",
                    "isDummyUser",
                    "disableNotifications",
                    "linkedin",
                    "researchGate",
                    "twitter",
                    "currentPosition",
                    "biography",
                    "createdAt",
                    "updatedAt",
                    "avatarId"
                ]
            },
            "SendMessageDto": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000
                    }
                },
                "required": ["message"]
            },
            "CreateWorkspaceDto": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "visibility": { "type": "object" },
                    "tenantId": { "type": "number" }
                },
                "required": ["name", "description", "visibility", "tenantId"]
            },
...
        }
    }
}